feat: separates GQL queries from main user test func
